Create base provider class View
