+++ packages/@uppy/provider-views/src/SearchProviderView/SearchProviderView.js
@@ -1,34 +1,25 @@
 const { h } = require('preact')
-const generateFileID = require('@uppy/utils/lib/generateFileID')
-const getFileType = require('@uppy/utils/lib/getFileType')
-const isPreviewSupported = require('@uppy/utils/lib/isPreviewSupported')
 const SearchInput = require('./InputView')
 const Browser = require('../Browser')
 const LoaderView = require('../Loader')
 const Header = require('./Header')
-const SharedHandler = require('../SharedHandler')
 const CloseWrapper = require('../CloseWrapper')
+const View = require('../View')
 
 /**
  * Class to easily generate generic views for Provider plugins
  */
-module.exports = class SearchProviderView {
+module.exports = class SearchProviderView extends View {
   static VERSION = require('../../package.json').version
 
-  #isHandlingScroll
-
   #searchTerm
-
-  #sharedHandler
 
   /**
    * @param {object} plugin instance of the plugin
    * @param {object} opts
    */
   constructor (plugin, opts) {
-    this.plugin = plugin
-    this.provider = opts.provider
-    this.#sharedHandler = new SharedHandler(plugin)
+    super(plugin, opts)
 
     // set default options
     const defaultOptions = {
@@ -45,17 +36,11 @@ module.exports = class SearchProviderView {
     this.search = this.search.bind(this)
     this.triggerSearchInput = this.triggerSearchInput.bind(this)
     this.addFile = this.addFile.bind(this)
-    this.preFirstRender = this.preFirstRender.bind(this)
-    this.handleError = this.handleError.bind(this)
     this.handleScroll = this.handleScroll.bind(this)
     this.donePicking = this.donePicking.bind(this)
-    this.cancelPicking = this.cancelPicking.bind(this)
-    this.clearSelection = this.clearSelection.bind(this)
 
     // Visual
     this.render = this.render.bind(this)
-
-    this.clearSelection()
 
     // Set default state for the plugin
     this.plugin.setPluginState({
@@ -65,6 +50,7 @@ module.exports = class SearchProviderView {
       directories: [],
       filterInput: '',
       isSearchVisible: false,
+      currentSelection: [],
     })
   }
 
@@ -79,15 +65,6 @@ module.exports = class SearchProviderView {
     this.plugin.setPluginState({ isInputMode: false, files })
   }
 
-  /**
-   * Called only the first time the provider view is rendered.
-   * Kind of like an init function.
-   */
-  preFirstRender () {
-    this.plugin.setPluginState({ didFirstRender: true })
-    this.plugin.onFirstRender()
-  }
-
   search (query) {
     if (query && query === this.#searchTerm) {
       // no need to search again as this is the same as the previous search
@@ -95,7 +72,7 @@ module.exports = class SearchProviderView {
       return
     }
 
-    return this.#sharedHandler.loaderWrapper(
+    return this.sharedHandler.loaderWrapper(
       this.provider.search(query),
       (res) => {
         this.#updateFilesAndInputMode(res, [])
@@ -108,72 +85,22 @@ module.exports = class SearchProviderView {
     this.plugin.setPluginState({ isInputMode: true })
   }
 
-  // @todo this function should really be a function of the plugin and not the view.
-  // maybe we should consider creating a base ProviderPlugin class that has this method
-  addFile (file) {
-    const tagFile = {
-      id: this.providerFileToId(file),
-      source: this.plugin.id,
-      data: file,
-      name: file.name || file.id,
-      type: file.mimeType,
-      isRemote: true,
-      body: {
-        fileId: file.id,
-      },
-      remote: {
-        companionUrl: this.plugin.opts.companionUrl,
-        url: `${this.provider.fileUrl(file.requestPath)}`,
-        body: {
-          fileId: file.id,
-        },
-        providerOptions: { ...this.provider.opts, provider: null },
-      },
-    }
-
-    const fileType = getFileType(tagFile)
-    // TODO Should we just always use the thumbnail URL if it exists?
-    if (fileType && isPreviewSupported(fileType)) {
-      tagFile.preview = file.thumbnail
-    }
-    this.plugin.uppy.log('Adding remote file')
-    try {
-      this.plugin.uppy.addFile(tagFile)
-    } catch (err) {
-      if (!err.isRestriction) {
-        this.plugin.uppy.log(err)
-      }
-    }
-  }
-
-  providerFileToId (file) {
-    return generateFileID({
-      data: file,
-      name: file.name || file.id,
-      type: file.mimeType,
-    })
-  }
-
-  handleError (error) {
-    const { uppy } = this.plugin
-    uppy.log(error.toString())
-    const message = uppy.i18n('companionError')
-    uppy.info({ message, details: error.toString() }, 'error', 5000)
-  }
-
-  handleScroll (e) {
-    const scrollPos = e.target.scrollHeight - (e.target.scrollTop + e.target.offsetHeight)
+  async handleScroll (event) {
     const query = this.nextPageQuery || null
 
-    if (scrollPos < 50 && query && !this.#isHandlingScroll) {
-      this.provider.search(this.#searchTerm, query)
-        .then((res) => {
-          const { files } = this.plugin.getPluginState()
-          this.#updateFilesAndInputMode(res, files)
-        }).catch(this.handleError)
-        .then(() => { this.#isHandlingScroll = false }) // always called
+    if (this.shouldHandleScroll(event) && query) {
+      this.isHandlingScroll = true
 
-      this.#isHandlingScroll = true
+      try {
+        const response = await this.provider.search(this.#searchTerm, query)
+        const { files } = this.plugin.getPluginState()
+
+        this.#updateFilesAndInputMode(response, files)
+      } catch (error) {
+        this.handleError(error)
+      } finally {
+        this.isHandlingScroll = false
+      }
     }
   }
 
@@ -181,20 +108,9 @@ module.exports = class SearchProviderView {
     const { currentSelection } = this.plugin.getPluginState()
     const promises = currentSelection.map((file) => this.addFile(file))
 
-    this.#sharedHandler.loaderWrapper(Promise.all(promises), () => {
+    this.sharedHandler.loaderWrapper(Promise.all(promises), () => {
       this.clearSelection()
     }, () => {})
-  }
-
-  cancelPicking () {
-    this.clearSelection()
-
-    const dashboard = this.plugin.uppy.getPlugin('Dashboard')
-    if (dashboard) dashboard.hideAllPanels()
-  }
-
-  clearSelection () {
-    this.plugin.setPluginState({ currentSelection: [] })
   }
 
   render (state, viewOptions = {}) {
@@ -206,7 +122,7 @@ module.exports = class SearchProviderView {
 
     const targetViewOptions = { ...this.opts, ...viewOptions }
     const { files, folders, filterInput, loading, currentSelection } = this.plugin.getPluginState()
-    const { isChecked, toggleCheckbox, filterItems } = this.#sharedHandler
+    const { isChecked, toggleCheckbox, filterItems } = this.sharedHandler
     const hasInput = filterInput !== ''
 
     const browserProps = {
